Moving attrs usage into a dedicated directory.
Some cleaning also.
